+++ bravado_core/model.py
@@ -87,6 +87,11 @@ def create_model_type(swagger_spec, model_name, model_spec):
     doc = docstring_property(partial(
         create_model_docstring, swagger_spec, model_spec))
 
+    def create(cls, kwargs):
+        self = cls.__new__(cls)
+        model_constructor(self, model_spec, swagger_spec, kwargs)
+        return self
+
     methods = dict(
         __doc__=doc,
         __eq__=lambda self, other: compare(self, other),
@@ -96,6 +101,7 @@ def create_model_type(swagger_spec, model_name, model_spec):
         __repr__=lambda self: create_model_repr(self, model_spec,
                                                 swagger_spec),
         __dir__=lambda self: model_dir(self, model_spec, swagger_spec),
+        create=classmethod(create),
         marshal=lambda self: marshal_model(swagger_spec, model_spec, self),
         unmarshal=staticmethod(lambda val: unmarshal_model(swagger_spec, model_spec, val)),
     )